Data row upload: add limit upload chunk size in bytes
