+++ libs/labelbox/src/labelbox/schema/internal/datarow_upload_constants.py
@@ -1,4 +1,5 @@
 MAX_DATAROW_PER_API_OPERATION = 150_000
 FILE_UPLOAD_THREAD_COUNT = 20
 UPSERT_CHUNK_SIZE = 10_000
+UPSERT_CHUNK_SIZE_BYTES = 10_000_000
 DOWNLOAD_RESULT_PAGE_SIZE = 5_000
